fix: 重复退出

+++ controller/accountController.js
@@ -57,9 +57,7 @@ exports.login = async (req, res) => {
     }
     if (user.dataValues.username === username && user.dataValues.password === password) {
       const userId = user.dataValues.userId
-      const token = jwt.sign({ id: userId }, JWT_KEY, { expiresIn: '2 days' })
-   
-     
+      const token = jwt.sign({ id: userId }, JWT_KEY, { expiresIn: '2 days' })   
       client.exists(userId).then(reply=>{
         if (reply === 1) {
           res.status(500).send('用户已登录');
@@ -85,33 +83,6 @@ exports.login = async (req, res) => {
         console.error(err);
         res.status(500);
       });
-      // client.connect().then(() => {
-      //   return client.exists(userId);
-      // }).then(reply => {
-      //   if (reply === 1) {
-      //     res.status(500).send('用户已登录');
-      //     throw new Error('User Already Logged In');
-      //   } else {
-      //     return client.set(userId, token);
-      //   }
-      // }).then(() => {
-      //   return User.findOne({
-      //     where: {
-      //       userId
-      //     }
-      //   });
-      // }).then(userInfo => {
-      //   res.status(200).send({
-      //     code: 200,
-      //     data: {
-      //       userInfo,
-      //       token,
-      //     }
-      //   });
-      // }).catch(err => {
-      //   console.error(err);
-      //   res.status(500);
-      // });
     } else {
       res.status(400).send('密码错误')
     }
@@ -125,13 +96,28 @@ exports.logout = (req, res) => {
     })
     return
   }
-  client.del(userId).then(()=>{
-    res.status(200).send({
-      message: '退出成功'
-    })
-  }).catch(()=>{
-    res.status(500);
+  console.log();
+  if(!client.get(userId)){
+   
+    return
+  }
+  client.get(userId).then(token=>{
+    console.log(token);
+    if(!token){
+      res.status(500).send({
+        message: '用户未登录'
+      })
+    }else {
+      client.del(userId).then(()=>{
+        res.status(200).send({
+          message: '退出成功'
+        })
+      }).catch(()=>{
+        res.status(500);
+      })
+    }
   })
+ 
  
 }
